requirements.txt
# PawID — Python Dependencies
# Install with: pip install -r requirements.txt

flask==3.0.0
tensorflow==2.13.0
numpy==1.24.3
Pillow==10.0.0
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
tqdm==4.66.1
werkzeug==3.0.1
python-dotenv
gunicorn